% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_petroff.R
\name{theme_petroff}
\alias{theme_petroff}
\title{A theme for petroff lab}
\usage{
theme_petroff(base_family = "sans", title_size = 22, facet_text_size = 14,
  facet_fill = "#C0C0C0", fill = "transparent", title_position = 0.5,
  axis_title_size = 14, axis_tick_size = 16, face = "bold",
  text_color = "black")
}
\arguments{
\item{base_family}{The font family to be used, defaults to Avenir}

\item{title_size}{Title size, defaults to 22}

\item{facet_text_size}{Facet text size, defaults to 12}

\item{facet_fill}{Color of facet rectangle, defaults to "#7fcdbb"}

\item{fill}{Background element rectangles, defaults to "transparent"}

\item{title_position}{Title position, defaults to 0.5}

\item{axis_title_size}{Axis title size, defaults to 14}

\item{axis_tick_size}{Size of ticks on axes, defaults to 12}

\item{face}{Typeface parameter, defaults to bold}

\item{text_color}{Color of text, defaults to black}

\item{colour}{Colour option for key, defaults to "transparent"}
}
\value{
A customized plot with Soo's theme!
}
\description{
A theme for petroff lab
}
